Added user management

+++ models/user.py
@@ -1,6 +1,6 @@
 from config import db
 
 class User(db.Model):
-    id = db.Column(db.SmallInteger, autoincrement=True, primary_key=True)
+    id = db.Column(db.Integer, autoincrement=True, primary_key=True)
     username = db.Column(db.String(50), nullable=False)
     password = db.Column(db.String(60), nullable=False)
